botlib: C++ compatibility for be_ai_move

+++ code/botlib/be_ai_move.c
@@ -621,7 +621,7 @@ static float DistanceFromLineSquared(vec3_t p, vec3_t lp1, vec3_t lp2)
 			(proj[j] < lp1[j] && proj[j] < lp2[j]))
 			break;
 	if (j < 3) {
-		if (fabs(proj[j] - lp1[j]) < fabs(proj[j] - lp2[j]))
+		if (fabs((double)(proj[j] - lp1[j])) < fabs((double)(proj[j] - lp2[j])))
 			VectorSubtract(p, lp1, dir);
 		else
 			VectorSubtract(p, lp2, dir);
@@ -1286,7 +1286,7 @@ static void BotCheckBlocked(bot_movestate_t *ms, vec3_t dir, int checkbottom, bo
 	//test for entities obstructing the bot's path
 	AAS_PresenceTypeBoundingBox(ms->presencetype, mins, maxs);
 	//
-	if (fabs(DotProduct(dir, up)) < 0.7)
+	if (fabs((double)(DotProduct(dir, up))) < 0.7)
 	{
 		mins[2] += sv_maxstep->value; //if the bot can step on
 		maxs[2] -= 10; //a little lower to avoid low ceiling
@@ -2058,7 +2058,7 @@ static bot_moveresult_t BotTravel_Elevator(bot_movestate_t *ms, aas_reachability
 		botimport.Print(PRT_MESSAGE, "bot on elevator\n");
 #endif //DEBUG_ELEVATOR
 		//if vertically not too far from the end point
-		if (fabs(ms->origin[2] - reach->end[2]) < sv_maxbarrier->value)
+		if (fabs((double)(ms->origin[2] - reach->end[2])) < sv_maxbarrier->value)
 		{
 #ifdef DEBUG_ELEVATOR
 			botimport.Print(PRT_MESSAGE, "bot moving to end\n");
@@ -2206,7 +2206,7 @@ static bot_moveresult_t BotFinishTravel_Elevator(bot_movestate_t *ms, aas_reacha
 	//
 	VectorSubtract(reach->end, ms->origin, topdir);
 	//
-	if (fabs(bottomdir[2]) < fabs(topdir[2]))
+	if (fabs((double)(bottomdir[2])) < fabs((double)(topdir[2])))
 	{
 		VectorNormalize(bottomdir);
 		EA_Move(ms->client, bottomdir, 300);
@@ -2653,8 +2653,8 @@ static bot_moveresult_t BotTravel_Grapple(bot_movestate_t *ms, aas_reachability_
 		result.flags |= MOVERESULT_MOVEMENTVIEW;
 		//
 		if (dist < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 2 &&
-			fabs(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 2)
+			fabs((double)(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0]))) < 2 &&
+			fabs((double)(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1]))) < 2)
 		{
 #ifdef DEBUG_GRAPPLE
 			botimport.Print(PRT_MESSAGE, "BotTravel_Grapple: activating grapple\n");
@@ -2721,8 +2721,8 @@ static bot_moveresult_t BotTravel_RocketJump(bot_movestate_t *ms, aas_reachabili
 	result.ideal_viewangles[PITCH] = 90;
 	//
 	if (dist < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 5)
+			fabs((double)(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0]))) < 5 &&
+			fabs((double)(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1]))) < 5)
 	{
 		//botimport.Print(PRT_MESSAGE, "between jump start and run start point\n");
 		hordir[0] = reach->end[0] - ms->origin[0];
@@ -2781,8 +2781,8 @@ static bot_moveresult_t BotTravel_BFGJump(bot_movestate_t *ms, aas_reachability_
 	dist = VectorNormalize(hordir);
 	//
 	if (dist < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 5)
+			fabs((double)(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0]))) < 5 &&
+			fabs((double)(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1]))) < 5)
 	{
 		//botimport.Print(PRT_MESSAGE, "between jump start and run start point\n");
 		hordir[0] = reach->end[0] - ms->origin[0];
